+++ lib/http2-proxy.js
@@ -1,15 +1,15 @@
 (function () {
-    var ConfEmitter = require('./h2-utils').ConfEmitter;
-    var define = require('./h2-utils').define;
-    var getOrigin = require('./h2-utils').getOrigin;
-    var resolvePort = require('./h2-utils').resolvePort;
-    var parseUrl = require('./h2-utils').parseUrl;
+    var ConfEmitter = require('./utils').ConfEmitter;
+    var define = require('./utils').define;
+    var getOrigin = require('./utils').getOrigin;
+    var resolvePort = require('./utils').resolvePort;
+    var parseUrl = require('./utils').parseUrl;
 
     var http2 = require('http2');
     var websocket = require('websocket-stream');
 
-    var Cache = require('./Cache.js').Cache;
-    var RequestInfo = require('./Cache.js').RequestInfo;
+    var Cache = require('./cache.js').Cache;
+    var RequestInfo = require('./cache.js').RequestInfo;
 
     // vars
     var activeWSConnections = {};
